+++ packages/polaris-viz/src/components/SimpleBarChart/stories/WithTrendIndicators.stories.tsx
@@ -41,6 +41,7 @@ WithTrendIndicators.args = {
       ],
       metadata: {
         trends: {
+          0: {},
           2: {
             value: '10%',
           },