+++ packages/polaris-viz/src/components/SimpleBarChart/utilities.ts
@@ -21,10 +21,6 @@ export function getLongestTrendIndicator(
       for (const [index, trend] of trendEntries) {
         const dataPoint = seriesData[index];
 
-        if (trend.value == null) {
-          continue;
-        }
-
         if (dataPoint?.value === highestPositive) {
           longestTrendIndicator.positive = estimateTrendIndicatorWidth(
             trend.value,